move ParameterSweep class to its own file
